### 2025-06-30 #8463 @kyleniemeyer
Hi @xzz105 @helgee @cwentland0 👋 are any of you available to review this submission for JOSS? The package is DARE: A MATLAB Toolbox for Design and Analysis of Ramjet/scramjet Engines. Thanks!
