### 2025-07-02 #8533 @editorialbot
[REVIEW]: DARE: A MATLAB Toolbox for Design and Analysis of Ramjet/scramjet Engines
**Submitting author:** @boraorcun (<a href="http://orcid.org/0000-0002-3577-4730">Bora Orcun Cakir</a>)
**Repository:** https://github.com/boraorcun/DARE.git
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
