+++ paper/paper.bib
@@ -174,7 +174,8 @@ doi = {10.2514/3.21349}
   number={2},
   pages={433--445},
   year={2013},
-  publisher={American Institute of Aeronautics and Astronautics}
+  publisher={American Institute of Aeronautics and Astronautics}.
+  doi = {10.2514/1.B34496}
 }
 
 @inbook{bhatia1990,
@@ -192,7 +193,7 @@ volume = {199},
 pages = {259-276},
 year = {2022},
 issn = {0094-5765},
-doi = {https://doi.org/10.1016/j.actaastro.2022.07.037},
+doi = {10.1016/j.actaastro.2022.07.037},
 url = {https://www.sciencedirect.com/science/article/pii/S0094576522003861},
 author = {Bora O. Cakir and Ali Can Ispir and Bayindir H. Saracoglu}
 }
@@ -204,7 +205,7 @@ volume = {217},
 pages = {349-362},
 year = {2024},
 issn = {0094-5765},
-doi = {https://doi.org/10.1016/j.actaastro.2024.01.036},
+doi = {10.1016/j.actaastro.2024.01.036},
 url = {https://www.sciencedirect.com/science/article/pii/S0094576524000444},
 author = {Ali Can Ispir and Bora O. Cakir and Bayindir H. Saracoglu}
 }
@@ -216,7 +217,7 @@ volume = {157},
 pages = {109811},
 year = {2025},
 issn = {1270-9638},
-doi = {https://doi.org/10.1016/j.ast.2024.109811},
+doi = {10.1016/j.ast.2024.109811},
 url = {https://www.sciencedirect.com/science/article/pii/S1270963824009404},
 author = {Bora O. Cakir and Ali Can Ispir}
 }
@@ -245,7 +246,8 @@ title = {Design space exploration for a scramjet engine by using data mining and
   number={3},
   pages={687--697},
   year={2016},
-  publisher={American Institute of Aeronautics and Astronautics}
+  publisher={American Institute of Aeronautics and Astronautics},
+  doi = {10.2514/1.B35887}
 }
 
 @misc{cvode,
@@ -256,13 +258,6 @@ title = {Design space exploration for a scramjet engine by using data mining and
   note         = {Accessed: 2024-12-21}
 }
 
-@phdthesis{fernandez2013simulation,
-  title={Simulation, design and analysis of air-breathing combined-cycle engines for high speed propulsion},
-  author={Fern{\'a}ndez Villac{\'e}, V{\'\i}ctor},
-  year={2013},
-  school={Aeronauticos}
-}
-
 @article{ispir2023reduced,
   title={Reduced-order modeling of supersonic fuel--air mixing in a multi-strut injection scramjet engine using machine learning techniques},
   author={Ispir, Ali C and Zdyba{\l}, Kamila and Saracoglu, Bayindir H and Magin, Thierry and Parente, Alessandro and Coussement, Axel},
@@ -270,7 +265,8 @@ title = {Design space exploration for a scramjet engine by using data mining and
   volume={202},
   pages={564--584},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.actaastro.2022.11.013}
 }
 
 @article{jachimowski1984chemical,
@@ -281,5 +277,6 @@ title = {Design space exploration for a scramjet engine by using data mining and
   number={2},
   pages={213--224},
   year={1984},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/0010-2180(84)90029-4}
 }
