+++ paper/paper.bib
@@ -174,7 +174,7 @@ doi = {10.2514/3.21349}
   number={2},
   pages={433--445},
   year={2013},
-  publisher={American Institute of Aeronautics and Astronautics}.
+  publisher={American Institute of Aeronautics and Astronautics},
   doi = {10.2514/1.B34496}
 }
 
