### 2025-08-29 #8533 @ecisneros8
Hello @kyleniemeyer, I'm conducting my review and have some additional comments beyond the checklist. However, I can't find the procedure to address such issues in the guidelines (my bad). Do I drop them here? Thanks in advance.
