### 2025-08-29 #8533 @kyleniemeyer
@ecisneros8 yes, you can just leave them here - no formal procedure needed. You can also open issues in the software repo and link/mention this review (so that they appear here) or vice versa.
