### 2025-09-18 #8533 @ecisneros8
@kyleniemeyer and @boraorcun I have attached a one-page review on DARE. I can update the checklist once the concerns stated in the review have been addressed. Please let me know if anything needs clarification.

[DARE-review.pdf](https://github.com/user-attachments/files/22410161/DARE-review.pdf)
