### 2025-11-04 9f21fdb
Update paper.md
The content added at line 38 addresses the major reviewer comment. Minor modifications have also been made to respond to the remaining reviewer remarks.
