+++ paper/paper.md
@@ -33,16 +33,18 @@ Advancing supersonic civil aviation with a focus on technical and environmental 
 
 # Statement of need
 
-Since the 1960s, researchers have developed and evaluated reduced-order models for high-speed propulsion systems, including ramjet and scramjet engines. Early models concentrated on simulating fuel injection and mixing effects, using a two-shock approach to analyze flow through intake contours [@bauer1966]. These efforts investigated critical combustion chamber parameters such as pre-ignition conditions, fuel mixing, and ignition techniques, while also assessing propulsion performance across varying altitudes and speeds. Later advancements introduced numerical methods that integrated finite-difference solutions with stirred reactor models and finite-rate chemistry, emphasizing design aspects like overall engine performance and weight optimization [@harsha1978; @edelman1981]. Subsequently, one-dimensional models incorporating Eulerian-Lagrangian frameworks were developed to study stable and unstable combustion modes and their influence on cycle performance [@bhatia1990]. These models also evaluated thrust losses caused by incomplete combustion and entropy generated from irreversibility [@riggins1995]. Some approaches combined low-fidelity propulsion analysis with structural integrity considerations for hypersonic engines, enabling performance assessments under aeroelastic conditions [@chavez1994; @bolender2007]. Despite these advancements, combustion modeling remained a limitation. To address this, @torrez2011 introduced a reduced-order engine model for ramjet and scramjet mixing and combustion, further enhancing the MASIV tool with the Shapiro method to predict thermal choking positions [@torrez2013new].
+Since the 1960s, researchers have developed and evaluated reduced-order models for high-speed propulsion systems, including ramjet and scramjet engines. Early models concentrated on simulating fuel injection and mixing effects, using a two-shock approach to analyze flow through intake contours [@bauer1966]. These efforts investigated critical combustion chamber parameters such as pre-ignition conditions, fuel mixing, and ignition techniques, while also assessing propulsion performance across varying altitudes and speeds. Later advancements introduced numerical methods that integrated finite-difference solutions with stirred reactor models and finite-rate chemistry, emphasizing design aspects like overall engine performance and weight optimization [@harsha1978; @edelman1981]. Subsequently, one-dimensional models incorporating Eulerian-Lagrangian frameworks were developed to study stable and unstable combustion modes and their influence on cycle performance [@bhatia1990]. These models also evaluated thrust losses caused by incomplete combustion and entropy generated from irreversibility [@riggins1995]. Some approaches combined low-fidelity propulsion analysis with structural integrity considerations for hypersonic engines, enabling performance assessments under aeroelastic conditions [@chavez1994; @bolender2007]. Despite these advancements, combustion modeling remained a limitation. To address this, @torrez2011 introduced a reduced-order engine model for ramjet and scramjet mixing and combustion, further enhancing the MASIV (Michigan Air Force Scramjet In Vehicle) tool with the Shapiro method to predict thermal choking positions [@torrez2013new].
+
+In parallel with these domain-specific developments, several modern frameworks employing one-dimensional modeling techniques have emerged for propulsion and system-level modeling. NASA’s OpenMDAO [@gray2019] provides a flexible, open-source Python framework for multidisciplinary design, analysis, and optimization, enabling coupling between aerodynamic, structural, and propulsion models. The Numerical Propulsion System Simulation (NPSS) [@lytle2000] offers a modular, extensible environment for thermodynamic cycle simulation of gas turbine engines, while tools such as SUAVE [@gawehn2020] and PyCycle [@gray2020pycycle] extend this approach to open-source aircraft and engine analysis. Within the MATLAB/Simulink ecosystem, NASA’s Toolbox for the Modeling and Analysis of Thermodynamic Systems (T-MATS) [@tmats] provides a component-based library for general thermodynamic and turbomachinery systems.
 
[…]
