### 2025-11-04 dba86f5
Update paper.bib
Some citations for the major paper review have been added and the citation for CVODE solver has been corrected.
