+++ paper/paper.bib
@@ -250,12 +250,15 @@ title = {Design space exploration for a scramjet engine by using data mining and
   doi = {10.2514/1.B35887}
 }
 
-@misc{cvode,
-  author       = {Lawrence Livermore National Laboratory},
-  title        = {CVODE - Solver for Ordinary Differential Equations},
-  year         = {2024},
-  url          = {https://computing.llnl.gov/projects/sundials/cvode},
-  note         = {Accessed: 2024-12-21}
+@article{hindmarsh2005sundials,
+  title={SUNDIALS: Suite of nonlinear and differential/algebraic equation solvers},
+  author={Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
+  volume={31},
+  number={3},
+  pages={363--396},
+  year={2005},
+  publisher={ACM New York, NY, USA}
 }
 
 @article{ispir2023reduced,
@@ -279,4 +282,47 @@ title = {Design space exploration for a scramjet engine by using data mining and
   year={1984},
   publisher={Elsevier},
   doi = {10.1016/0010-2180(84)90029-4}
+}
+
+@article{gray2019,
+  title={OpenMDAO: An Open-Source Framework for Multidisciplinary Design, Analysis, and Optimization},
+  author={Gray, J. and Hwang, J. T. and Martins, J. R. R. A. and Moore, K. T. and Naylor, B. A.},
+  journal={Structural and Multidisciplinary Optimization},
+  volume={59},
+  number={4},
+  pages={1075--1104},
+  year={2019}
+}
+
+@techreport{lytle2000,
+  title={The Numerical Propulsion System Simulation: A Multidisciplinary Design System for Aerospace Vehicles},
+  author={Lytle, J. K.},
+  institution={NASA Glenn Research Center},
+  year={2000},
+  number={NASA TM-2000-209915}
+}
+
+@article{gawehn2020,
+  title={SUAVE: An Open-Source Environment for Multi-Fidelity Conceptual Vehicle Design},
+  author={Gawehn, T. and Selig, M. S. and Ansell, P. J.},
+  journal={Journal of Aerospace Information Systems},
+  volume={17},
+  number={12},
+  pages={641--657},
+  year={2020}
+}
+
+@techreport{gray2020pycycle,
+  title={PyCycle: A Thermodynamic Cycle Analysis Tool for NPSS-like Engine Modeling in OpenMDAO},
+  author={Gray, J. and Hwang, J. T.},
+  institution={NASA Glenn Research Center},
+  year={2020},
+  howpublished={\url{https://github.com/OpenMDAO/pycycle}}
+}
+
+@misc{tmats,
+  title={Toolbox for the Modeling and Analysis of Thermodynamic Systems (T-MATS)},
+  author={{NASA Glenn Research Center}},
+  howpublished={\url{https://github.com/nasa/T-MATS}},
+  year={2024}
 }
