### 2025-11-04 325b5b0
Update paper.md
CVODE citation in Line 74 has been removed.
