+++ paper/paper.md
@@ -44,7 +44,7 @@ The third module addresses nozzle design and analysis, simulating flow expansion
 
 # Research application
 
-`DARE` has been primarily used within the EU funded Stratofly and More\&Less projects to focus on the design, analysis, and optimization of ramjet and dual-mode ramjet/scramjet propulsion systems, which are critical for supersonic and hypersonic vehicles operating beyond Mach 3 (@zhang2016quasi; @cvode; @cakir2022]. DARE provided assessments for flow development, combustion performance, and overall propulsive characteristics such as thrust, fuel consumption, specific impulse, and efficiency are analyzed under varying conditions, including flight Mach number, altitude, intake geometry, and fuel-air equivalence ratio [@ispir2024; @cakir2024]. Sensitivity analyses, including Shapley Additive Explanations (SHAP) and feature importance studies, identify key design parameters influencing engine performance. These methodologies are demonstrated to efficiently explore the design space, generate performance maps, and quantify propulsion system feasibility for high-supersonic cruise vehicles at an affordable computational cost. `DARE` has also been used in conjunction with machine learning techniques to support data-driven modeling of high-speed propulsion systems. In these studies, artificial neural networks were applied to both operational and design variables of ramjet and dual-mode ramjet/scramjet engines to enable a comprehensive exploration of the design space. The resulting machine learning models were trained to predict and represent the propulsive performance of these engines with improved computational efficiency [@ispir2022; @cakir2023]. Furthermore, efforts were made to enhance the capabilities of DARE, including the addition of a fuel-air mixing model in the combustor. This was achieved by modeling the mixing efficiency as a function of fuel injector configuration parameters along the scramjet propulsion system duct axis, using machine learning applied to a CFD database [@ispir2023reduced]. 
+`DARE` has been primarily used within the EU funded Stratofly and More\&Less projects to focus on the design, analysis, and optimization of ramjet and dual-mode ramjet/scramjet propulsion systems, which are critical for supersonic and hypersonic vehicles operating beyond Mach 3 (@zhang2016quasi; @cakir2022]. DARE provided assessments for flow development, combustion performance, and overall propulsive characteristics such as thrust, fuel consumption, specific impulse, and efficiency are analyzed under varying conditions, including flight Mach number, altitude, intake geometry, and fuel-air equivalence ratio [@ispir2024; @cakir2024]. Sensitivity analyses, including Shapley Additive Explanations (SHAP) and feature importance studies, identify key design parameters influencing engine performance. These methodologies are demonstrated to efficiently explore the design space, generate performance maps, and quantify propulsion system feasibility for high-supersonic cruise vehicles at an affordable computational cost. `DARE` has also been used in conjunction with machine learning techniques to support data-driven modeling of high-speed propulsion systems. In these studies, artificial neural networks were applied to both operational and design variables of ramjet and dual-mode ramjet/scramjet engines to enable a comprehensive exploration of the design space. The resulting machine learning models were trained to predict and represent the propulsive performance of these engines with improved computational efficiency [@ispir2022; @cakir2023]. Furthermore, efforts were made to enhance the capabilities of DARE, including the addition of a fuel-air mixing model in the combustor. This was achieved by modeling the mixing efficiency as a function of fuel injector configuration parameters along the scramjet propulsion system duct axis, using machine learning applied to a CFD database [@ispir2023reduced]. 
  
 # Acknowledgements
 
