### 2025-11-04 528c7d9
Update paper.bib
A citation correction.
