+++ paper/paper.bib
@@ -250,17 +250,6 @@ title = {Design space exploration for a scramjet engine by using data mining and
   doi = {10.2514/1.B35887}
 }
 
-@article{hindmarsh2005sundials,
-  title={SUNDIALS: Suite of nonlinear and differential/algebraic equation solvers},
-  author={Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S},
-  journal={ACM Transactions on Mathematical Software (TOMS)},
-  volume={31},
-  number={3},
-  pages={363--396},
-  year={2005},
-  publisher={ACM New York, NY, USA}
-}
-
 @article{ispir2023reduced,
   title={Reduced-order modeling of supersonic fuel--air mixing in a multi-strut injection scramjet engine using machine learning techniques},
   author={Ispir, Ali C and Zdyba{\l}, Kamila and Saracoglu, Bayindir H and Magin, Thierry and Parente, Alessandro and Coussement, Axel},
