### 2025-11-05 f2a66ad
Update paper.md
