+++ paper/paper.md
@@ -39,7 +39,7 @@ In parallel with these domain-specific developments, several modern frameworks e
 
 Numerous low-fidelity design and analysis studies have been conducted to characterize the performance of ramjet engines. However, most of these tools focus on individual propulsion system components rather than providing a holistic methodology. To date, there has been no significant effort to integrate intake design approaches with combustion analysis modules, and only a limited number of studies consider the combined effects of flight conditions and design parameters across the entire propulsion flow path. While analyzing the performance of individual components is crucial for understanding the physical phenomena that influence various design considerations for ramjet and scramjet engines, a thorough exploration of the design envelope is essential. This approach ensures an accurate definition of mission requirements and enables the effective optimization of design choices for high-speed aircraft.
 
-The `DARE` (A MATLAB Toolbox for Design and Analysis of Ramjet/scramjet Engines) tool is introduced as a design and analysis framework that integrates individual approaches for high-speed propulsion components into a unified low-fidelity methodology. This tool aims to enable cost-effective characterization of the design space for high-speed propulsion systems. The ramjet/scramjet propulsion flow path comprises four primary components: the intake, isolator, combustor, and nozzle. The proposed analytical tool provides a fully integrated flow path analysis divided into three main modules. The first module focuses on the intake, facilitating the necessary freestream flow modulation before entering the isolator. For ramjet configurations, this module incorporates a normal shock assumption. The resulting flow properties are then used in the second module, which models combustion. This module calculates flow evolution within the combustion chamber using one-dimensional steady, inviscid flow equations coupled with detailed chemical kinetics and JANAF tables, employing the SUNDIALS code [@hindmarsh2005], developed by Lawrence Livermore National Laboratory.
+The `DARE` (a toolbox for Design and Analysis of Ramjet/scramjet Engines) tool is introduced as a design and analysis framework that integrates individual approaches for high-speed propulsion components into a unified low-fidelity methodology. This tool aims to enable cost-effective characterization of the design space for high-speed propulsion systems. The ramjet/scramjet propulsion flow path comprises four primary components: the intake, isolator, combustor, and nozzle. The proposed analytical tool provides a fully integrated flow path analysis divided into three main modules. The first module focuses on the intake, facilitating the necessary freestream flow modulation before entering the isolator. For ramjet configurations, this module incorporates a normal shock assumption. The resulting flow properties are then used in the second module, which models combustion. This module calculates flow evolution within the combustion chamber using one-dimensional steady, inviscid flow equations coupled with detailed chemical kinetics and JANAF tables, employing the SUNDIALS code [@hindmarsh2005], developed by Lawrence Livermore National Laboratory.
 The third module addresses nozzle design and analysis, simulating flow expansion across various expansion ratios and nozzle geometries using one-dimensional steady, inviscid flow equations under cold flow conditions. Key performance metrics, including thrust, fuel consumption, and specific impulse, are then calculated to evaluate engine performance for each design.
 
 # Research application
