### 2025-11-10 713877d
Update nozzleModule.m
Modifications have been applied to the nozzle module to:
- Fix the errors occurring between lines 190 and 200.
- Remove input parameters related to nozzle design (these have been transferred to the main script, DARE.m).
