+++ DARE/nozzleModule.m
@@ -1,4 +1,4 @@
-function [nozzOut] = nozzleModule(combOut,nozzCond,combCond,atmCond)
+function [nozzOut] = nozzleModule(combOut,nozzCond,combCond,atmCond,choice,pressure_difference)
 
 global Ru engineMode
 
@@ -15,19 +15,8 @@ for m=2:1:row
     area_gradient(m,2) = area_profile_comb_nozzle(m,2) - area_profile_comb_nozzle(m-1,2);
 end
 
-
-nozzle_analysis_options = ["Expansion with the provided area profile", "Perfect expansion"];
-choice = menu('Select an analysis option:', nozzle_analysis_options);
 if choice == 1 && area_profile_comb_nozzle(end,1) < Comb_Length + nozzCond{2}
     error('The length info given in the area profile is not sufficient, please provide the area profile along with the full-lenght, update nozzle and combustor lengths, or select perfect expansion');
-end
-pressure_difference = input('Please enter the desired pressure difference (exit pressure - ambient pressure) in [%] of ambient pressure for designing the nozzle: ');
-if pressure_difference > 0
-    fprintf('You entered a pressure difference of %.2f Pa. This is for underexpanded nozzle design\n', pressure_difference);
-elseif pressure_difference < 0 && pressure_difference > -30
-    fprintf('You entered a pressure difference of %.2f Pa. This is for overexpanded nozzle design\n', pressure_difference);
-elseif pressure_difference <= -30
-    fprintf('For this overexpanded nozzle design, thrust production is not possible according to Sommerfield criteria');
 end
 
 %% CALLING THE CVODE SOLVER
@@ -61,10 +50,11 @@ if strcmp( engineMode, 'RAM' )
     imax = 120000000; % maximum iteration number
     xout = 1 + Comb_Length;
     x = combOut{2}; x0 = Comb_Length;
-    y = combOut{3}; y0 = y;
+    y = combOut{3}; 
+	y(2) = y(2) + 5; % ramjet solutions are valid only with ±5m/s velocity range at the throat of the combustor
+	y0 = y;
     module = 'nozzle';
     CVodeFree;
-    y(2) = y(2) + 5; % ramjet solutions are valid only with ±5m/s velocity range at the throat of the combustor
     CVodeInit(@ramjet_combustor_nozzle, 'BDF', 'Newton', x0, y0, options); % initializing the CVODE solver for the solution of ramjet combustor and nozzle modules
     if choice == 1 && nozzle_key == true
         while iout < imax
@@ -187,7 +177,7 @@ elseif strcmp( engineMode, 'SCRAM' )
         Cp_NO = specific_heat_calculation_fun(y(4),'NO',Ru,MW(11)); % specific heat of NO [kJ/kg/K]
         Cp_HNO = specific_heat_calculation_fun(y(4),'HNO',Ru,MW(12)); % specific heat of HNO [kJ/kg/K]
         Cp_NO2 = specific_heat_calculation_fun(y(4),'NO2',Ru,MW(13)); % specific heat of NO2 [kJ/kg/K]
[…]
