### 2025-11-10 d15351d
Update DARE.m
Modifications were applied to the main DARE code to include the nozzle input parameters directly within it, instead of defining them inside the nozzle module.
