### 2025-11-10 ea1facd
Update combustorModule.m
The plot showing the temperature rise throughout the numerical iterations has been removed from the code.
