### 2025-11-10 d123f14
Update reaction_rates_calculation_fun.m
Basic modifications were applied to the code to remove warning signals during calculations.
