### 2025-11-10 1603c50
Update specific_heat_calculation_fun.m
Basic modifications were applied to the code to remove warning signals during calculations.
