### 2025-11-10 0491585
Added a new folder and file to the examples
To automate the examples and simplify the comparison of performance outputs, tests/testDARE_Examples.m has been added.
