### 2025-11-10 4aae78d
Update DARE_RamExample1.m
The individual nozzle modules in each example folder have been removed, and the main DARE codes for all examples have been updated to accommodate the modifications to the nozzle module.
