### 2025-11-11 381a24b
Update DARE_RamExample2.m
The individual nozzle modules in each example folder have been removed, and the main DARE codes for all examples have been updated to accommodate the modifications to the nozzle module.
