### 2025-11-11 4b67668
Update DARE_RamExample1.m
