### 2025-11-11 65c711d
Update DARE_ScramExample1.m
The individual nozzle modules in each example folder have been removed, and the main codes for all examples have been updated to accommodate the modifications to the nozzle module.
