### 2025-11-11 10855b2
Update DARE_ScramExample2.m
The individual nozzle modules in each example folder have been removed, and the main codes for all examples have been updated to accommodate the modifications to the nozzle module.
