### 2025-11-11 c61352f
Update README.md
The README file has been updated to include information about the added files and folders used for automating the examples.
