### 2026-02-17 #8533 @kyleniemeyer
@boraorcun I wanted to check in on your submission—it's been a few months. Are you still revising this? At this point, unless you anticipate being done shortly, we should probably pull your submission, and then you can resubmit to JOSS when ready.
